remove rest
